requirements.txt
absl-py==1.0.0
alembic==1.7.4
altair==4.2.0
anyio==3.3.3
apache-airflow==2.2.0
apache-airflow-client==2.1.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bandit==1.7.0
bcj-cffi==0.5.1
beautifulsoup4==4.9.3
bleach==3.3.1
blinker==1.4
blis==0.7.7
Brotli==1.0.9
bs4==0.0.1
cachetools==5.0.0
cairocffi==1.3.0
catalogue==2.0.7
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.6
cfgv==3.3.0
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
click-default-group==1.2.2
click-plugins==1.1.1
clickclick==20.10.2
cligj==0.7.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
coverage==6.1.1
croniter==1.0.15
cryptography==3.4.8
cycler==0.10.0
cymem==2.0.6
decorator==5.0.6
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.2
dnspython==2.1.0
docutils==0.16
dunk==0.1.0a0
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.1.0
Faker==8.0.0
filelock==3.0.12
Fiona==1.8.20
flake8==3.9.2
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==2.0
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.2.0/fr_core_news_sm-3.2.0-py3-none-any.whl
fr-dep-news-trf @ https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.2.0/fr_dep_news_trf-3.2.0-py3-none-any.whl
gast==0.5.3
geopandas==0.9.0
ghp-import==2.0.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.17
grpcio==1.44.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
httpcore==0.13.7
httpx==0.19.0
huggingface-hub==0.5.1
identify==2.2.11
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
inflection==0.5.1
iniconfig==1.1.1
ipykernel==5.5.3
ipytest==0.11.0
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.16
isodate==0.6.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupytext==1.11.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
lazy-object-proxy==1.4.3
libclang==13.0.0
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
markdown-it-py==0.6.2
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.4.1
mccabe==0.6.1
mdit-py-plugins==0.2.6
mergedeep==1.3.4
missingno==0.4.2
mistune==0.8.4
mkdocs==1.2.3
mkdocs-autorefs==0.3.0
mkdocs-custommill==1.2.6
mkdocs-material==7.3.6
mkdocs-material-extensions==1.0.3
mkdocstrings==0.16.2
multivolumefile==0.2.3
munch==2.5.0
murmurhash==1.0.6
mypy==0.910
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.0
numpy==1.20.2
oauthlib==3.2.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.1
patsy==0.5.1
pbr==5.6.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.0.2
plotly==4.14.3
pluggy==0.13.1
pre-commit==2.13.0
preshed==3.0.6
prison==0.2.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.20.0
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
py7zr==0.16.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.2
pydeck==0.7.1
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==1.7.1
pymdown-extensions==9.0
Pympler==1.0.1
pyparsing==2.4.7
pyppmd==0.15.1
pyproj==3.1.0
pyrsistent==0.18.0
pytest==6.2.4
pytest-cov==3.0.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-markdown-math==0.8
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytkdocs==0.12.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
pyyaml_env_tag==0.1
pyzmq==22.0.3
pyzstd==0.14.4
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.1
retrying==1.3.3
rfc3986==1.5.0
rich==12.0.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
semver==2.13.0
Send2Trash==1.7.1
sentencepiece==0.1.96
setproctitle==1.2.2
Shapely==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.2.4
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-streamlit==1.0.3
spacy-transformers==1.1.5
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlite-fts4==1.0.1
sqlite-utils==3.26
srsly==2.4.3
statsmodels==0.12.2
stevedore==3.3.0
streamlit==1.8.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
texttable==1.6.4
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.15
threadpoolctl==2.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.11.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.17.0
typer==0.4.1
typing-extensions==3.10.0.0
tzdata==2022.1
tzlocal==4.2
unicodecsv==0.14.1
unidiff==0.7.3
urllib3==1.26.7
validators==0.18.2
virtualenv==20.6.0
wasabi==0.9.1
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.14.0
WTForms==2.3.3
XlsxWriter==3.0.1
zipp==3.6.0
